Initial implementation
Implement the initial functionality to read the CISA NVD JSON, save the
CVE IDs it contains to a MongoDB collection. Since we are using the Beanie
library to provide a MongoDB ODM we use Python's asyncio features for
the implementation. NOTE: This initial implementation isn't parsing the
URLs correctly. Will rebase once functionality is correct.