Add time for lambda, retry for URL, and ssm params
The nvdsync lambda is now functional but the ssm
parameters haven't officially been tested. The implementation
was copied from the kevsync lambda so it should be
good to go. Also added retries for the URL socket
connection. The default is set for 10 attempts with
5 second sleeps in between each attempt. Also the
lambda will timeout with the default 5 minute window.
Increased the time so the lambda can run until all
the CVEs are processed into the database.