Add task groups to enable concurrency
Currently this will crash when it hits
CVEs from 2021. I suspect the issue is
related to the event loop. That being
said the documents do seem to be
processing into the DB substantially
faster. Just need to make fixes so
the lambda doesn't crash or halt
due to over processing.